Support for RA PvD ID option and multiple RA per interface
This is a basic implementation of the RA PvD ID option:
draft-bruneau-intarea-provisioning-domains-01.
The option can be configured using the configuration options:
- pvd_id
- pvd_http
- pvd_legacy

The patch also adds the ability for multiple RAs to be sent on
the same interface: draft-ietf-rtgwg-enterprise-pa-multihoming-01
This was almost the case already, but router sollicitations and
rtnetlink messages had to be considered for all interfaces.